requirements.txt
phable==0.1.14
python-dotenv
pandas
scikit-learn
matplotlib
